Änderungen für den Fakultätsadmin

+++ htdocs/admin_institut.php
@@ -43,7 +43,6 @@ if ($RESOURCES_ENABLE) {
 	require_once ($RELATIVE_PATH_RESOURCES."/lib/ResourcesAssign.class.php");
 }
 	
-
 ###
 ### Submit Handler
 ###
@@ -77,8 +76,12 @@ while ( is_array($HTTP_POST_VARS)
   
   ## Create a new Institut
   case "create_x":
-    ## Do we have all necessary data?
-    if (empty($Name)) {
+  if (!$perm->is_fak_admin()){
+	  $msg = "error§<b>Sie haben nicht die Berechtigung, um neue Einrichtungen zu erstellen!</b>";
+	  break;
+  }
+  ## Do we have all necessary data?
+	if (empty($Name)) {
       $msg="error§<b>Bitte geben sie eine Bezeichnung f&uuml;r die Einrichtung ein!</b>";
       $i_view="new";
       break;
@@ -94,15 +97,18 @@ while ( is_array($HTTP_POST_VARS)
 
     ## Create an id
     $i_id=md5(uniqid($hash_secret));
-	## Namen der Fakultaet durch Fakultaets_id erstzen
-	$db2->query("SELECT * FROM Fakultaeten WHERE Name = '$Fakultaet'");
-  	if ($db2->next_record()) {
-		$Fakultaet = $db2->f("Fakultaets_id");
+	if (!$Fakultaet){
+		if ($perm->have_perm("root")){
+		$Fakultaet = $i_id;
+		} else {
+			$msg = "error§<b>Sie haben nicht die Berechtigung neue Fakult&auml;ten zu erstellen";
+			break;
+		}
 	}
 	## insert the Institut...
 	if ($home == "")
 		$home = "http://www.studip.de";
-    $query = "insert into Institute values('$i_id','$Name','$Fakultaet','$strasse','$plz', '$home', '$telefon', '$email', '$fax', '$type','".time()."', '".time()."')";
+    $query = "insert into Institute (Institut_id,Name,fakultaets_id,Strasse,Plz,url,telefon,email,fax,type,mkdate,chdate) values('$i_id','$Name','$Fakultaet','$strasse','$plz', '$home', '$telefon', '$email', '$fax', '$type','".time()."', '".time()."')";
     $db->query($query);
     if ($db->affected_rows() == 0) {
       $msg="error§<b>Datenbankoperation gescheitert: $query </b>";
@@ -117,23 +123,24 @@ while ( is_array($HTTP_POST_VARS)
    $i_view = $i_id;
 
    //This will select the new institute later for navisgation (=>links_admin.inc.php)
-   $admin_inst_id =$i_id; 
+   $admin_inst_id = $i_id;
+   openInst($i_id);
   break;
 
   ## Change Institut name
   case "i_edit_x":
 
-    ## Do we have all necessary data?
+  if (!$perm->have_studip_perm("admin",$i_id)){
+	$msg = "error§<b>Sie haben nicht die Berechtigung diese Einrichtungen zu ver&auml;ndern!</b>";
+	break;
+  }
+	  
+	## Do we have all necessary data?
     if (empty($Name)) {
       $msg="error§<b>Bitte geben sie eine Bezeichnug f&uuml;r die Einrichtung ein!</b>";
       break;
     }
 
-	## Namen der Fakultaet durch Fakultaets_id erstzen
-	$db2->query("SELECT * FROM Fakultaeten WHERE Name = '$Fakultaet'");
-  	if ($db2->next_record()) {
-		$Fakultaet = $db2->f("Fakultaets_id");
-	}
     ## Update Institut information.
     $query = "UPDATE Institute SET Name='$Name', fakultaets_id='$Fakultaet', Strasse='$strasse', Plz='$plz', url='$home', telefon='$telefon', fax='$fax', email='$email', type='$type' ,chdate=".time()." where Institut_id = '$i_id'";
     $db->query($query);
@@ -155,7 +162,17 @@ while ( is_array($HTTP_POST_VARS)
       $msg="error§<b>Diese Einrichtung kann nicht gel&ouml;scht werden, da noch Veranstaltungen an dieser Einrichtung existieren!</b>";
       break;
     }
-    
+	$db->query("SELECT a.Institut_id,a.Name, IF(a.Institut_id=a.fakultaets_id,1,0) AS is_fak, count(b.Institut_id) as num_inst FROM Institute a LEFT JOIN Institute b ON (a.Institut_id=b.fakultaets_id) WHERE a.Institut_id ='$i_id' AND b.Institut_id!='$i_id' AND a.Institut_id=a.fakultaets_id GROUP BY a.Institut_id ");
+	$db->next_record();
+	if($db->f("num_inst")){
+		$msg="error§<b>Diese Einrichtung kann nicht gel&ouml;scht werden, da sie den Status Fakult&auml;t hat, und noch andere Einrichtungen zugeordnet sind!</b>";
+		break;
+	}
+	if ($db->f("is_fak") && !$perm->have_perm("root")){
+		$msg="error§<b>Sie haben nicht die Berechtigung Fakult&auml;ten zu l&oumlschen!</b>";
+		break;
+	}
+		
     ## Delete that Institut.
     $query = "delete from Institute where Institut_id='$i_id'";
     $db->query($query);
@@ -275,45 +292,58 @@ if ($i_view=="delete") {
 	die;
 }
 	
-	
-if ($i_view)
-	{
-	$db->query("SELECT * FROM user_inst WHERE Institut_id ='$i_view' AND user_id = '$user->id' AND inst_perms = 'admin'");
-	IF ($db->next_record() OR $perm->have_perm("root"))
-{		
 
-		
-	
-  	if ($i_view<>"new")
-  		{
-		$db->query("SELECT Institute.*, count(Seminar_id) AS number FROM Institute LEFT OUTER JOIN seminare USING (Institut_id) WHERE Institute.Institut_id ='$i_view' GROUP BY seminare.Institut_id");
+if ($perm->have_studip_perm("admin",$i_view) || $i_view == "new") {
+
+	if ($i_view != "new") {
+		$db->query("SELECT a.*,b.Name AS fak_name, count(Seminar_id) AS number FROM institute a LEFT JOIN institute b ON (b.Institut_id=a.fakultaets_id) LEFT JOIN seminare c ON (a.Institut_id=c.Institut_id) WHERE a.Institut_id ='$i_view' GROUP BY a.Institut_id");
 		$db->next_record();
-		}
-
+		$db2->query("SELECT a.Institut_id,a.Name,count(b.Institut_id) as num_inst FROM institute a LEFT JOIN institute b ON (a.Institut_id=b.fakultaets_id) WHERE a.Institut_id ='$i_view' AND b.Institut_id!='$i_view' AND a.Institut_id=a.fakultaets_id GROUP BY a.Institut_id ");
+		$db2->next_record();
+		$_num_inst = $db2->f("num_inst");
+	}
 	$i_id= $db->f("Institut_id");
-  ?>
-  <tr><td class="blank" colspan=2>
-  <table border=0 align="center" width="50%" cellspacing=0 cellpadding=2>
+	?>
+<tr><td class="blank" colspan=2>
+<table border=0 align="center" width="50%" cellspacing=0 cellpadding=2>
 	<form method="POST" name="edit" action="<? echo $PHP_SELF?>">
-	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" >Name: </td><td class="<? echo $cssSw->getClass() ?>" ><input type="text" name="Name" size=32 maxlength=254 value="<?php echo htmlReady($db->f("Name")) ?>"></td></tr>
-	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" >Fakult&auml;t</td>
-		<td class="<? echo $cssSw->getClass() ?>" align=left><select name="Fakultaet">
+	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" >Name: </td><td class="<? echo $cssSw->getClass() ?>" ><input type="text" name="Name" size=50 maxlength=254 value="<?php echo htmlReady($db->f("Name")) ?>"></td></tr>
+	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" >Fakult&auml;t:</td>
+		<td class="<? echo $cssSw->getClass() ?>" align=left>
 		<?php
-		$db2->query("SELECT * FROM Fakultaeten ORDER BY Name");
-		while ($db2->next_record()) {
-			printf ("<option %s> %s</option>", $db2->f("Fakultaets_id") == $db->f("fakultaets_id") ? "selected" : "", htmlReady($db2->f("Name")));
+		if ($perm->is_fak_admin() && ($perm->have_studip_perm("admin",$db->f("fakultaets_id")) || $i_view == "new")){
+			if ($_num_inst){
+				echo "\n<font size=\"-1\"><b>Diese Einrichtung hat den Status einer Fakult&auml;t.<br>
+				Es wurden bereits $_num_inst andere Einrichtungen zugeordnet.</b></font>";
+				echo "\n<input type=\"hidden\" name=\"Fakultaet\" value=\"$i_id\">";
+			} else {
+				echo "\n<select name=\"Fakultaet\">";
+				if ($perm->have_perm("root")) {
+					printf ("<option %s value=\"%s\">Diese Einrichtung hat den Status einer Fakult&auml;t</option>", ($db->f("fakultaets_id") == $db->f("Institut_id")) ? "selected" : "", $db->f("Institut_id"));
+					$db2->query("SELECT Institut_id,Name FROM Institute WHERE Institut_id=fakultaets_id AND fakultaets_id !='". $db->f("institut_id") ."' ORDER BY Name");
+				} else {
+					$db2->query("SELECT a.Institut_id,Name FROM user_inst a LEFT JOIN Institute USING (Institut_id) WHERE user_id='$user->id' AND inst_perms='admin' AND a.Institut_id=fakultaets_id AND fakultaets_id !='". $db->f("institut_id") ."' ORDER BY Name");
+				}
+				while ($db2->next_record()) {
+					printf ("<option %s value=\"%s\"> %s</option>", ($db2->f("Institut_id") == $db->f("fakultaets_id"))  ? "selected" : "", $db2->f("Institut_id"),htmlReady($db2->f("Name")));
+				}
+				echo "</select>";
+			}
+		} else {
+			echo htmlReady($db->f("fak_name")) . "\n<input type=\"hidden\" name=\"Fakultaet\" value=\"" . $db->f("fakultaets_id") . "\">";
 		}
+			
 		?>
-	</select></td>
+	</td>
 	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" >Bezeichnung: </td><td class="<? echo $cssSw->getClass() ?>" ><select name="type">
 	<? 
 	$i=0;
 	foreach ($INST_TYPE as $a) {
 		$i++;
 		if ($i==$db->f("type"))
-			echo "<option selected value=$i>".$INST_TYPE[$i]["name"]."</option>";
+			echo "<option selected value=\"$i\">".$INST_TYPE[$i]["name"]."</option>";
 		else
-			echo "<option value=$i>".$INST_TYPE[$i]["name"]."</option>";		
+			echo "<option value=\"$i\">".$INST_TYPE[$i]["name"]."</option>";		
 		}
 	?></select></td></tr>
 	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" >Strasse: </td><td class="<? echo $cssSw->getClass() ?>" ><input type="text" name="strasse" size=32 maxlength=254 value="<?php echo htmlReady($db->f("Strasse")) ?>"></td></tr>
@@ -327,17 +357,17 @@ if ($i_view)
 	<tr <? $cssSw->switchClass() ?>><td class="<? echo $cssSw->getClass() ?>" colspan=2 align="center">
 	
 	<? 
-	if ($i_view<>"new")
+	if ($i_view != "new")
 		{
 		?>
 		<input type="hidden" name="i_id"   value="<?php $db->p("Institut_id") ?>">
 		<input type="IMAGE" name="i_edit" src="./pictures/buttons/uebernehmen-button.gif" border=0 value=" Ver&auml;ndern ">
 		<?
-		if ($db->f("number") < 1):
+		if ($db->f("number") < 1 && !$_num_inst){
 			?>
 			&nbsp;<input type="IMAGE" name="i_kill" src="./pictures/buttons/loeschen-button.gif" border=0 value=" L&ouml;schen ">
 			<?
-		endif;
+		}
 		}
 	else
 		{
@@ -348,37 +378,36 @@ if ($i_view)
 	</td></tr></table>
 	</form>
 	<br>
- 
-  	<?
-  	
-  	if ($i_view<>"new")
+<?
+	
+	if ($i_view != "new")
 		{
- 		$db->query("SELECT * FROM faecher LEFT  JOIN fach_inst USING (fach_id) WHERE institut_id = '$i_id'");
+		$db->query("SELECT * FROM faecher LEFT  JOIN fach_inst USING (fach_id) WHERE institut_id = '$i_id'");
 		$cssSw->resetClass();
- 		?>
- 		<table border=0 align="center" width="80%" cellspacing=0 cellpadding=2>
- 		<tr><td width="100%" colspan=2><br>&nbsp;Dieser Einrichtung sind folgende Studienf&auml;cher zugeordnet:<br><br></th></tr>
- 		<tr><th width="80%" align="center">Name</th><th width="20%" align="center">Aktion</th><tr>
+		?>
+		<table border=0 align="center" width="80%" cellspacing=0 cellpadding=2>
+		<tr><td width="100%" colspan=2><br>&nbsp;Dieser Einrichtung sind folgende Studienf&auml;cher zugeordnet:<br><br></th></tr>
+		<tr><th width="80%" align="center">Name</th><th width="20%" align="center">Aktion</th><tr>
 		<?
- 		while ($db->next_record()) {
- 			$cssSw->switchClass();
- 			echo"<tr><td class=\"".$cssSw->getClass()."\">", htmlReady($db->f("name")), "</td><td class=\"".$cssSw->getClass()."\" align=\"center\"><form method=\"POST\" name=\"kill_f\" action=", $PHP_SELF, "><input type=\"IMAGE\" name=\"kill_fach\" src=\"./pictures/buttons/entfernen-button.gif\" border=0 value=\" Zuordnung aufheben\"><input type=\"hidden\" name=\"i_view\" value=\"", $i_id, "\"><input type=\"hidden\" name=\"fach_id\" value=\"", $db->f("fach_id"),"\"></td></form></tr>";
- 		}
- 		$cssSw->switchClass();
- 		echo"<tr><td class=\"".$cssSw->getClass()."\"><form method=\"POST\" name=\"add_f\" action=", $PHP_SELF, "><select name=\"fach_id\" size=1>";
- 		$db2->query("SELECT * FROM faecher ORDER BY name");
- 		while ($db2->next_record())
- 			{
- 			$ftmp = $db2->f("fach_id");
- 			$db->query("SELECT * FROM fach_inst WHERE institut_id = '$i_view' AND fach_id = '$ftmp'");
-	 		IF (!$db->next_record())
-	 			echo "<option value=".$db2->f("fach_id").">", htmlReady(substr($db2->f("name"),0,80));
- 			}
- 		echo "</select></td><td class=\"".$cssSw->getClass()."\" align=\"center\"><input type=\"IMAGE\" src=\"./pictures/buttons/zuordnen-button.gif\" border=0 name=\"add_fach\" value=\" Zuordnen\"><input type=\"hidden\" name=\"i_view\" value=\"", $i_id, "\"></td></form></tr>";
- 		echo "</table><br><br>";
-  		}
-  	}
+		while ($db->next_record()) {
+			$cssSw->switchClass();
+			echo"<tr><td class=\"".$cssSw->getClass()."\">", htmlReady($db->f("name")), "</td><td class=\"".$cssSw->getClass()."\" align=\"center\"><form method=\"POST\" name=\"kill_f\" action=", $PHP_SELF, "><input type=\"IMAGE\" name=\"kill_fach\" src=\"./pictures/buttons/entfernen-button.gif\" border=0 value=\" Zuordnung aufheben\"><input type=\"hidden\" name=\"i_view\" value=\"", $i_id, "\"><input type=\"hidden\" name=\"fach_id\" value=\"", $db->f("fach_id"),"\"></td></form></tr>";
+		}
+		$cssSw->switchClass();
+		echo"<tr><td class=\"".$cssSw->getClass()."\"><form method=\"POST\" name=\"add_f\" action=", $PHP_SELF, "><select name=\"fach_id\" size=1>";
+		$db2->query("SELECT * FROM faecher ORDER BY name");
+		while ($db2->next_record())
+			{
+			$ftmp = $db2->f("fach_id");
+			$db->query("SELECT * FROM fach_inst WHERE institut_id = '$i_view' AND fach_id = '$ftmp'");
+			IF (!$db->next_record())
+				echo "<option value=".$db2->f("fach_id").">", htmlReady(substr($db2->f("name"),0,80));
+			}
+		echo "</select></td><td class=\"".$cssSw->getClass()."\" align=\"center\"><input type=\"IMAGE\" src=\"./pictures/buttons/zuordnen-button.gif\" border=0 name=\"add_fach\" value=\" Zuordnen\"><input type=\"hidden\" name=\"i_view\" value=\"", $i_id, "\"></td></form></tr>";
+		echo "</table><br><br>";
+		}
 }
+
 
 echo"</table>";
 page_close();
